Struct emit::pipeline::reference::PipelineRef

pub struct PipelineRef {
    // some fields omitted
}

PipelineRef is "mouth" of the pipeline, into which events are fed.

Methods

impl PipelineRef

fn new(head: Box<Emit + Sync>, level: LogLevelFilter) -> PipelineRef

fn is_enabled(&self, level: LogLevel) -> bool

Check if the specified log level is enabled for the pipeline.

fn emit(&self, event: Event<'static>)

Emit an event through the pipeline. Code wishing to conditionally emit events based on the level should call is_enabled() first.